tools/encounter-simulator/requirements.txt
# LangChain and LLM dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Data validation
pydantic>=2.0.0

# OpenAI API
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Terminal formatting
rich>=13.0.0
